requirements.txt
fastapi
uvicorn[standard]
python-multipart
pydantic
pydantic-settings
python-dotenv
openai
requests
pillow
numpy
